DolphinQt: Move wiimote controller ui logic to its own widget
