+++ Source/Core/DolphinQt/Config/ControllersWindow.h
@@ -19,6 +19,7 @@ class QGroupBox;
 class QLabel;
 class QPushButton;
 class QRadioButton;
+class WiimoteControllersWidget;
 
 class ControllersWindow final : public QDialog
 {
@@ -27,18 +28,11 @@ public:
   explicit ControllersWindow(QWidget* parent);
 
 private:
-  void OnWiimoteModeChanged();
-  void UpdateDisabledWiimoteControls();
   void OnGCTypeChanged(int state);
   void SaveSettings();
-  void OnBluetoothPassthroughSyncPressed();
-  void OnBluetoothPassthroughResetPressed();
-  void OnWiimoteRefreshPressed();
   void OnGCPadConfigure();
-  void OnWiimoteConfigure();
 
   void CreateGamecubeLayout();
-  void CreateWiimoteLayout();
   void CreateMainLayout();
   void ConnectWidgets();
   void LoadSettings();
@@ -54,23 +48,7 @@ private:
   std::array<QHBoxLayout*, 4> m_gc_groups;
 
   // Wii Remote
-  QGroupBox* m_wiimote_box;
-  QGridLayout* m_wiimote_layout;
-  std::array<QLabel*, 4> m_wiimote_labels;
-  std::array<QComboBox*, 4> m_wiimote_boxes;
-  std::array<QPushButton*, 4> m_wiimote_buttons;
-  std::array<QHBoxLayout*, 4> m_wiimote_groups;
-  std::array<QLabel*, 2> m_wiimote_pt_labels;
-
-  QRadioButton* m_wiimote_emu;
-  QRadioButton* m_wiimote_passthrough;
-  QPushButton* m_wiimote_sync;
-  QPushButton* m_wiimote_reset;
-  QCheckBox* m_wiimote_continuous_scanning;
-  QCheckBox* m_wiimote_real_balance_board;
-  QCheckBox* m_wiimote_speaker_data;
-  QCheckBox* m_wiimote_ciface;
-  QPushButton* m_wiimote_refresh;
+  WiimoteControllersWidget* m_wiimote_controllers;
 
   // Common
   CommonControllersWidget* m_common;
